requirements.txt
connexion == 2.6.0
Flask-cors >= 3.0.10
python_dateutil == 2.6.0
requests >= 2.26.0
selenium >= 4.0.0
setuptools >= 21.0.0
